pyproject.toml
[project]
name = "quotes_price"
version = "0.0.0"
description = "A validator that checks if a price quote in an unexpected currency is present in the text"
authors = [
    {name = "Cartesia AI", email = "join@cartesia.ai"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "guardrails-ai>=0.3.2"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]